TES3::ItemVariables: There's a duplicate structure here that needs to be removed. But, for now, make it at least use the right variable type for charge.

+++ MWSE/TES3Inventory.h
@@ -12,7 +12,7 @@ namespace TES3 {
 		void * lockData; // 0x8
 		int condition; // 0xC
 		union {
-			int charge;
+			float charge;
 			Actor* soul;
 		} enchantData; // 0x10
 		TES3::Script * script; // 0x14
